requirements.txt
adbutils==0.6.2
aiohttp==3.5.4
allure-pytest==2.6.4
allure-python-commons==2.6.4
Appium-Python-Client==0.47
appnope==0.1.0
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
ddt==1.2.1
decorator==4.4.0
deepdiff==4.0.7
Deprecated==1.2.6
deprecation==2.0.7
enum34==1.1.6
fabric==2.4.0
fake==0.8
Faker==1.0.7
humanize==0.5.1
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
invoke==1.2.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.13.3
jsonpickle==1.2
logzero==1.5.0
lxml==4.3.3
more-itertools==7.0.0
moviepy==1.0.0
multidict==4.5.2
namedlist==1.7
numpy==1.16.3
ordered-set==3.1.1
packaging==19.2
parameterized==0.7.0
paramiko==2.5.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
proglog==0.1.9
progress==1.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==4.5.0
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==7.0
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
retry==0.9.2
selenium==3.141.0
six==1.12.0
text-unidecode==1.2
tornado==3.2
tqdm==4.32.1
traitlets==4.3.2
uiautomator2==1.3.6
urllib3==1.25.2
wcwidth==0.1.7
whichcraft==0.6.1
wrapt==1.11.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yarl==1.3.0
yml==0.0.1
